+++ src/systems/system_sets/mod.rs
@@ -1,0 +1,7 @@
+use bevy::prelude::*;
+
+#[derive(SystemSet, Clone, Hash, Debug, PartialEq, Eq)]
+pub struct MovementSet;
+
+#[derive(SystemSet, Clone, Hash, Debug, PartialEq, Eq)]
+pub struct AttackSet;