+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Migrated core, temp, ui, world and ship systems to systems/mod.rs file for better overview
 - Move set_window_icon temporary system to PreStartup schedule
 - Move configure_physics_engine to GameCreate state
+- Move system sets to new systems module
 
 ## [0.0.19] - 2023-10-31